requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
discord.py==2.5.2
feedparser==6.0.11
forex-python==1.9.2
frozenlist==1.7.0
idna==3.10
lxml==5.4.0
multidict==6.4.4
propcache==0.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.4
sgmllib3k==1.0.0
simplejson==3.20.1
six==1.17.0
soupsieve==2.7
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
yarl==1.20.1
